TXNG-21: Added ability to serialize/deserialize transaction attempts.
Motivation
----------
In the interest of N1QL transactions support, we are adding the ability to
generically serialize and deserialize a transaction attempt such that it
can be shipped to another service (such as N1QL).

Changes
-------
Added methods to serialize a transaction attempt and additionally added
methods which take the serialized data and create a new transaction object.
Additionally we cleaned up the SerializeContext code which is not used.