Add Peek to priority queue
Replace delay queue with priority queue. Difference being that
it does not check if the item is "ready" during pop, due to
how heap.Pop works this was causing items in the queue to be
incorrectly reordered. Instead of checking ready during pop we
now have a Peek function which returns the item at head which can
be checked for readyness before calling Pop.