TXNG-113: Fixed some fetch failure edge cases.
When fetch fails due to a CasMismatch, we need to fast-fail the replace
unstaging when forceWrite=true so we don't write incorrect data.
Additionally added some additional guarding and context to the same case
on the insert path (previously would have failed with a BadValue server
error due to trying to insert with no content).

Change-Id: Iac7ab15c616ba4eeee76f22a4898052efcd27aef
Reviewed-on: http://review.couchbase.org/c/gocbcore-transactions/+/150022
Tested-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>